triton-server-prebuilt: add recipe
The triton-server recipe does not yet support grpc builds.

It also may be useful to have a comparison of the prebuilt content,
in the case issues are found when attempting to run the triton-server
built from source.

This recipe works, including with grpc, after a couple of hacks
to pull out libraries which aren't supported due to library
version mismatches, and to symlink in a libb64 reference which
expects a different name.

A backport of this onto scarthgap should also work with minor
modifications to use the 2.49 release.

Signed-off-by: Dan Walkes <danwalkes@trellis-logic.com>